@@ -5,13 +5,13 @@ plugins {
 }
 
 android {
-    compileSdkVersion 30
-    buildToolsVersion "30.0.2"
+    compileSdkVersion 32
+    buildToolsVersion "30.0.3"
 
     defaultConfig {
-        applicationId "com.example.tictoctoe"
+        applicationId "com.recogniseerror.cartoontictoc"
         minSdkVersion 18
-        targetSdkVersion 30
+        targetSdkVersion 32
         versionCode 1
         versionName "1.0"
 
@@ -47,24 +47,25 @@ dependencies {
     testImplementation 'junit:junit:4.13.2'
     androidTestImplementation 'androidx.test.ext:junit:1.1.3'
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.4.0'
-    // Google Service
-    implementation 'com.google.android.gms:play-services-auth:20.1.0'
-
-    // For Show Image
-    implementation 'com.squareup.picasso:picasso:2.71828'
 
     // For Multiple language
     implementation 'com.android.support:multidex:1.0.3'
 
+    implementation 'com.github.dpaukov:combinatoricslib3:3.3.2'
+
+   /* // Google Service
+    //implementation 'com.google.android.gms:play-services-auth:20.1.0'
+
+    // For Show Image
+    implementation 'com.squareup.picasso:picasso:2.71828'
+
     // For GIF
-    implementation 'pl.droidsonroids.gif:android-gif-drawable:1.2.19'
+    //implementation 'pl.droidsonroids.gif:android-gif-drawable:1.2.19'
 
     //Lifecycle
     implementation 'androidx.lifecycle:lifecycle-extensions:2.2.0'
 
-    implementation 'com.github.dpaukov:combinatoricslib3:3.3.2'
-
     implementation 'com.github.bumptech.glide:glide:4.12.0'
-    annotationProcessor 'com.github.bumptech.glide:compiler:4.12.0'
+    annotationProcessor 'com.github.bumptech.glide:compiler:4.12.0'*/
 
 }
