@@ -1,4 +1,4 @@
-package com.example.tictoctoe;
+package com.recogniseerror.cartoontictoc;
 
 import androidx.appcompat.app.AppCompatActivity;
 
@@ -6,7 +6,7 @@ import android.content.Intent;
 import android.os.Bundle;
 import android.os.Handler;
 
-import com.example.tictoctoe.Game.GameHome;
+import com.recogniseerror.cartoontictoc.Game.GameHome;
 
 public class MainActivity extends AppCompatActivity {
 
@@ -15,14 +15,11 @@ public class MainActivity extends AppCompatActivity {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
 
-        new Handler().postDelayed(new Runnable() {
-            @Override
-            public void run() {
-                Intent goToMain;
-                goToMain = new Intent(MainActivity.this, GameHome.class);
-                startActivity(goToMain);
-                finish();
-            }
+        new Handler().postDelayed(() -> {
+            Intent goToMain;
+            goToMain = new Intent(MainActivity.this, GameHome.class);
+            startActivity(goToMain);
+            finish();
         },1200);
     }
 }
