@@ -1,3 +1,3 @@
-package com.example.tictoctoe
+package com.recogniseerror.cartoontictoc
 
 data class Cell(val i: Int, val j: Int)
