@@ -1,4 +1,4 @@
-package com.example.tictoctoe;
+package com.recogniseerror.cartoontictoc;
 
 import org.junit.Test;
 
